requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
torch>=2.0.0
transformers>=4.35.0
supabase>=2.0.0
PyJWT>=2.8.0

